+++ test/reviewer-round-end.test.js
@@ -138,6 +138,7 @@ function mkFixture() {
     getUiSettings: () => ({ get: () => ({}) }),
     getPersistence: () => persistence,
     getTemplates: () => ({ list: () => [SHIPPED_REVIEWER_TEMPLATE] }),
+    listAllTemplates: () => [SHIPPED_REVIEWER_TEMPLATE],
     notifyOS: () => {},
     intentEnabled,
     withoutPrivilegedIntentsFor: require('../intent-registry').withoutPrivilegedIntentsFor,