+++ test/ticket-rework-reasons.test.js
@@ -124,6 +124,7 @@ function mkFixture() {
     getUiSettings: () => ({ get: () => ({}) }),
     getPersistence: () => persistence,
     getTemplates: () => ({ list: () => [SHIPPED_REVIEWER_TEMPLATE] }),
+    listAllTemplates: () => [SHIPPED_REVIEWER_TEMPLATE],
     notifyOS: () => {},
     intentEnabled,
     withoutPrivilegedIntentsFor: require('../intent-registry').withoutPrivilegedIntentsFor,